belajar tipe data di python

+++ Pertemuan 3/CobaVar.java
@@ -5,6 +5,6 @@ public class CobaVar {
         double tinggiBadan = 155;
         System.out.println("Nama Saya " + namaDepan);
         System.out.println("Umur saya " + umur);
-        System.out.println("Tinggi badan saya " + tinggiBadan);
+        System.out.println("Tinggi badan saya " + (int) tinggiBadan);
     }
 }
